requirements.txt
absl-py==1.4.0
accelerate==0.21.0
aiohttp==3.8.4
aiosignal==1.3.1
annoy==1.17.3
appdirs==1.4.4
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
boltons==23.0.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.4
contourpy==1.1.0
cycler==0.11.0
dask==2023.9.0
debugpy==1.6.7.post1
denoising-diffusion-pytorch==1.8.6
docker-pycreds==0.4.0
einops==0.6.1
ema-pytorch==0.2.3
exceptiongroup==1.1.3
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
frozenlist==1.3.3
fsspec==2023.6.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
gluonts==0.13.2
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.2
h5py==3.9.0
holidays==0.26
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
iniconfig==2.0.0
Jinja2==3.1.2
joblib==1.2.0
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.6
lightning-utilities==0.8.0
lit==16.0.6
llvmlite==0.40.1
locket==1.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.57.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pacmap==0.7.0
pandas==1.5.3
partd==1.4.0
pathtools==0.1.2
Pillow==9.5.0
plotly==5.15.0
pluggy==1.3.0
protobuf==4.23.3
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.9
pyparsing==3.0.9
pytest==7.4.1
python-dateutil==2.8.2
pytorch-fid==0.3.0
pytorch-lightning==2.0.3
pytz==2023.3
PyYAML==6.0
pyzmq==19.0.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.25.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.47
stribor==0.2.0
sympy==1.12
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.33.0
termcolor==2.3.0
tf-slim==1.1.0
threadpoolctl==3.2.0
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchcde==0.2.5
torchdiffeq==0.2.2
torchmetrics==0.11.4
torchsde==0.2.5
torchtyping==0.1.4
torchvision==0.15.2
tqdm==4.65.0
trampoline==0.1.2
triton==2.0.0
typeguard==4.1.3
typing_extensions==4.7.1
ujson==5.8.0
urllib3==1.26.16
wandb==0.15.12
Werkzeug==2.3.6
wrapt==1.15.0
xgboost==1.7.6
yarl==1.9.2
zipp==3.15.0
